requirements.txt
# django
openai
langchain
transformers
serpapi
streamlit
gradio
requests
kaggle
graphviz
matplotlib
langchain_community
beautifulsoup4
python-dotenv
langchain-community
langchain-openai
forefront
djangorestframework